remove autoload.js script tag
