+++ index.html
@@ -11,7 +11,6 @@
 <body>
   <div id="app" class="dark:dark-bg"></div>
   <script type="module" src="/src/main.ts"></script>
-  <script src="assets/Live2d/autoload.js"></script>
 </body>
 
 </html>
