+++ src/App.vue
@@ -4,15 +4,15 @@
  * @Email: kanoqwq@qq.com
  * @Date: 2023-04-17 14:47:15
  * @Last Modified by: kanoqwq
- * @Last Modified time: 2023-04-17 14:52:24
+ * @Last Modified time: 2023-04-19 17:48:30
  * @Description: Description
  */
 import { ref, onMounted } from 'vue';
 import Search from './components/Search.vue'
 // iconfont
 import '@/assets/js/iconfont.js'
+import Live2D from "@/components/Live2D/index.vue"
 import useStore from './store'
-
 const Configs = useStore.Configs();
 let background = Configs.getBackgroundImage(-1)
 let main = ref<HTMLDivElement>()
@@ -27,6 +27,7 @@ onMounted(() => {
 <template>
   <div class="main" ref="main">
     <Search />
+    <Live2D/>
   </div>
 </template>
