+++ src/components/Settings.vue
@@ -6,13 +6,29 @@
                 <div class="background-set flex">
                     <span class="m-3">背景图片URL:</span>
                     <input class="rounded mr-1 input" type="text" v-model="backgroundImage">
-                    <button
-                        class="rounded p-2  transition-all bg-gray-600 hover:bg-purple-700 focus:outline-none focus:ring-2 focus:ring-purple-600 focus:ring-opacity-50"
-                        @click="modifyBg">修改</button>
-                    <button
-                        class="rounded p-2 ml-1 transition-all bg-gray-600 hover:bg-purple-700 focus:outline-none focus:ring-2 focus:ring-purple-600 focus:ring-opacity-50"
-                        @click="resetBg">重置</button>
+                    <div class="options">
+                        <button
+                            class="rounded p-2  transition-all bg-gray-600 hover:bg-purple-700 focus:outline-none focus:ring-2 focus:ring-purple-600 focus:ring-opacity-50"
+                            @click="modifyBg">修改</button>
+                        <button
+                            class="rounded p-2 ml-1 transition-all bg-gray-600 hover:bg-purple-700 focus:outline-none focus:ring-2 focus:ring-purple-600 focus:ring-opacity-50"
+                            @click="resetBg">重置</button>
+                    </div>
+
                 </div>
+                <div class="background-set flex">
+                    <span class="m-3">背景图片URL:</span>
+                    <input class="rounded mr-1 input" type="text" v-model="backgroundImage">
+                    <div class="options">
+                        <button
+                            class="rounded p-2  transition-all bg-gray-600 hover:bg-purple-700 focus:outline-none focus:ring-2 focus:ring-purple-600 focus:ring-opacity-50"
+                            @click="modifyBg">修改</button>
+                        <button
+                            class="rounded p-2 ml-1 transition-all bg-gray-600 hover:bg-purple-700 focus:outline-none focus:ring-2 focus:ring-purple-600 focus:ring-opacity-50"
+                            @click="resetBg">重置</button>
+                    </div>
+                </div>
+
             </div>
         </div>
     </Transition>
@@ -102,6 +118,10 @@ const resetBg = () => {
 
     .model {
         position: fixed;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        flex-direction: column;
         // min-width: 500px;
         left: 50%;
         top: 50%;
@@ -111,18 +131,21 @@ const resetBg = () => {
         background-color: rgba(255, 255, 255, .3);
         color: #000;
         border-radius: 10px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
+
         box-shadow: 0px 0px 15px #000;
 
         .background-set {
             height: 40px;
             overflow: hidden;
+            margin: 10px 0;
 
             .input {
                 background-color: rgb(31 41 55 / .5);
                 outline: none;
+            }
+
+            button {
+                height: 100%;
             }
         }
     }
